rhino/requirements.txt
ConfigParser==3.5.0
django==1.11.6
ConfigParser==3.5.0
avro-python3==1.8.2
certifi==2017.7.27.1
chardet==3.0.4
fastavro==0.14.10
idna==2.6
requests==2.18.4
urllib3==1.22
confluent-kafka==0.11.0
django-rest-framework==0.1.0
mysqlclient==1.3.12
django_crontab==0.7.1
pymongo==3.5.1
pymysql==0.7.11
selenium==3.7.0
celery
django-celery-results
django-celery-beat
threadpool
xlwt
xlrd
xlutils
flower
django-cors-headers
